chore: update cosmic-panel-config and friends...
this commit updates cosmic-panel-config, which now has size_wings and size_center fields. these fields can be used to override the size of applets on any of the panel sides. the templates are updated to account for this, setting these values to None for cosmic, mac, and ubuntu, and for windows, it sets size_wings to Some((Some(XS), Some(XS))). if we don't want this, I can change those values to None as well, but I feel that it looks more windows-like to have those values set smaller than the app tray.

Signed-off-by: Ryan Brue <ryanbrue.dev@gmail.com>